emacs: Set 'eww' as the default web browser
Because this was nested within the use-package definition of eww, when
attempting to open a link you would see the warning that there are no
browsers to use. This is especially true when using emacs-nox or
compiling emacs without X. To correct this I've moved the browser
definition to the 'Startup' section.