emacs: suggest mdempsky/gocode rather than stamblerre/gocode
I was using stamblerre/gocode which is supposed to support Go modules,
but it generates a ton of errors which were causing the creation of a
TON of buffers in emacs. Switching to mdempsky/gocode seems to work
fine, even in projects that use Go modules and it doesn't generate
dozens of errors, which trigger go-eldoc to generate buffers.

Anyway, go-eldoc is now enabled by default. My replication test of
moving the cursor over a series of functions quickly enough for the
context errors to be generated with the stamblerre version, but as
mentioned above this doesn't happen with the upstream version.

So, what is the Go modules "support" needed for?