emacs: disable everything I don't commonly use
Using 'use-package's `:disabled` option, it is possible to disable any
use-package definition. A large amount of what I have in my config is
highly related to the current situation I'm developing around, but
I've added a lot of things I _might_ use over time. With this change,
the existing configuration does not need to be modified, and if I need
to use some specific functionality I can remove the `:disabled` line
from whatever I need to start using.

This is part of the 'Conditional Loading' section of the use-package
README:

- https://github.com/jwiegley/use-package#conditional-loading